move passing tests using Simulacron from Broken
Moved some tests from BasicTypeTests, TupleTests and ClusterSimulacronTests. None of the tests were modified.
