+++ src/Cassandra.BrokenIntegrationTests/Core/TupleTests.cs
@@ -33,94 +33,6 @@ namespace Cassandra.IntegrationTests.Core
         private const string TableName = "users_tuples";
 
         [Test]
-        public void DecodeTupleValuesSingleTest()
-        {
-            TestCluster.PrimeFluent(
-                b => b.WhenQuery("SELECT * FROM " + TableName + " WHERE id = 1")
-                      .ThenRowsSuccess(
-                          new[]
-                          {
-                              ("id", DataType.Int),
-                              ("phone", DataType.Tuple(DataType.Text, DataType.Text, DataType.Int)),
-                              ("achievements", DataType.List(DataType.Tuple(DataType.Text, DataType.Int)))
-                          },
-                          r => r.WithRow(1, new Tuple<string, string, int>("home", "1234556", 1), null)));
-
-            var row = Session.Execute("SELECT * FROM " + TableName + " WHERE id = 1").First();
-            var phone1 = row.GetValue<Tuple<string, string, int>>("phone");
-            var phone2 = row.GetValue<Tuple<string, string, int>>("phone");
-            Assert.IsNotNull(phone1);
-            Assert.IsNotNull(phone2);
-            Assert.AreEqual("home", phone1.Item1);
-            Assert.AreEqual("1234556", phone1.Item2);
-            Assert.AreEqual(1, phone1.Item3);
-        }
-
-        [Test]
-        public void DecodeTupleNullValuesSingleTest()
-        {
-            TestCluster.PrimeFluent(
-                b => b.WhenQuery("SELECT * FROM " + TableName + " WHERE id = 11")
-                      .ThenRowsSuccess(
-                          new[]
-                          {
-                              ("id", DataType.Int),
-                              ("phone", DataType.Tuple(DataType.Text, DataType.Text, DataType.Int)),
-                              ("achievements", DataType.List(DataType.Tuple(DataType.Text, DataType.Int)))
-                          },
-                          r => r.WithRow(11, new Tuple<string, string, int?>("MOBILE", null, null), null)));
-            var row = Session.Execute("SELECT * FROM " + TableName + " WHERE id = 11").First();
-            var phone = row.GetValue<Tuple<string, string, int>>("phone");
-            Assert.IsNotNull(phone);
-            Assert.AreEqual("MOBILE", phone.Item1);
-            Assert.AreEqual(null, phone.Item2);
-            Assert.AreEqual(0, phone.Item3);
-
-            TestCluster.PrimeFluent(
-                b => b.WhenQuery("SELECT * FROM " + TableName + " WHERE id = 12")
-                      .ThenRowsSuccess(
-                          new[]
-                          {
-                              ("id", DataType.Int),
-                              ("phone", DataType.Tuple(DataType.Text, DataType.Text, DataType.Int)),
-                              ("achievements", DataType.List(DataType.Tuple(DataType.Text, DataType.Int)))
-                          },
-                          r => r.WithRow(12, new Tuple<string, string, int?>(null, "1222345", null), null)));
-            row = Session.Execute("SELECT * FROM " + TableName + " WHERE id = 12").First();
-            phone = row.GetValue<Tuple<string, string, int>>("phone");
-            Assert.IsNotNull(phone);
-            Assert.AreEqual(null, phone.Item1);
-            Assert.AreEqual("1222345", phone.Item2);
-            Assert.AreEqual(0, phone.Item3);
-        }
-
-        [Test]
-        public void DecodeTupleAsNestedTest()
-        {
-            TestCluster.PrimeFluent(
-                b => b.WhenQuery("SELECT * FROM " + TableName + " WHERE id = 21")
-                      .ThenRowsSuccess(
-                          new[]
-                          {
-                              ("id", DataType.Int),
-                              ("phone", DataType.Tuple(DataType.Text, DataType.Text, DataType.Int)),
-                              ("achievements", DataType.List(DataType.Tuple(DataType.Text, DataType.Int)))
-                          },
-                          r => r.WithRow(
-                              21,
-                              null,
-                              new List<Tuple<string, int?>>
-                              {
-                                  new Tuple<string, int?>("Tenacious", 100),
-                                  new Tuple<string, int?>("Altruist", 12)
-                              })));
-            var row = Session.Execute("SELECT * FROM " + TableName + " WHERE id = 21").First();
-
-            var achievements = row.GetValue<List<Tuple<string, int>>>("achievements");
-            Assert.IsNotNull(achievements);
-        }
-
-        [Test]
         public void EncodeDecodeTupleAsNestedTest()
         {
             var achievements = new List<Tuple<string, int>>
